+++ app/src/main/java/com/luna/budgetapp/domain/model/CategoryTotalProjection.kt
@@ -2,5 +2,5 @@ package com.luna.budgetapp.domain.model
 
 data class CategoryTotalProjection(
     val category: Category,
-    val total: Double
+    val total: Long
 )
